feat: Added default pcb and yaml file choosers
This commit adds default file choosers for kicad_pcb and yaml files.
The default behaviour for KiCAD is to use the project name
as the directory and the individual file names.
This can be used to make a reasonable first guess for the
kicad_pcb file name. If however the file is not found, the first
kicad_pcb file found in the current working directory is used.
The same is applied for the yaml file, in addition to a _pinion suffix.